Fuzz: add fake aead impl to fuzzlib
